InventarioProject/requirements.txt
asgiref
boto3
botocore
dj-database-url
Django
django-annoying
django-appconf
django-compressor
django-debug-toolbar
django-import-export
django-environ
django-filter
django-redis
django-storages
dotenv
djangorestframework
et-xmlfile
gunicorn
jmespath
openpyxl
packaging
pandas
pillow
psycopg2-binary
python-dateutil
rcssmin
rjsmin
s3transfer
six
sqlparse
typing_extensions
tzdata
urllib3
whitenoise